requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
pandas==2.1.4
pydantic==2.11.9
pydantic-settings==2.10.1
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.1
python-dotenv==1.0.0
